integrated remote testing onto the django test suite

+++ test/test_ssh_to_remote.bash
@@ -15,29 +15,30 @@ REMOTE_USER="$1"
 REMOTE_HOST="$2"
 SECRET_KEY="$3"
 BACKUP_FOLDER="$4"
-
-REMOTE_USER="mpavez"
-REMOTE_HOST="localhost"
-SECRET_KEY="/home/mpavez/.ssh/id_rsa"
-BACKUP_FOLDER="/home/mpavez/bkps/test"
-
+# REMOTE_USER="mpavez"
+# REMOTE_HOST="localhost"
+# SECRET_KEY="/home/mpavez/.ssh/id_rsa"
+# BACKUP_FOLDER="/home/mpavez/bkps/test"
 
 # settings
-echo "################################################################################"
-echo "# test ssh connection to remote visualization server"
-echo "################################################################################"
-echo "---------"
-echo "using REMOTE_USER: $REMOTE_USER"
-echo "using REMOTE_HOST: $REMOTE_HOST"
-echo "using SECRET_KEY: $SECRET_KEY"
-echo "using BACKUP_FOLDER: $BACKUP_FOLDER"
-echo "---------"
+echo ""
+echo " AndroidRequestBackups test: run ssh script on remote server"
+echo " -----------------------------------------------------------------------------"
+if [ -z "$REMOTE_USER"   ]; then (>&2 echo " >>> (FAIL) Required parameter REMOTE_USER"  ) ; exit 1; fi
+if [ -z "$REMOTE_HOST"   ]; then (>&2 echo " >>> (FAIL) Required parameter REMOTE_HOST"  ) ; exit 1; fi
+if [ -z "$SECRET_KEY"    ]; then (>&2 echo " >>> (FAIL) Required parameter SECRET_KEY"   ) ; exit 1; fi
+if [ -z "$BACKUP_FOLDER" ]; then (>&2 echo " >>> (FAIL) Required parameter BACKUP_FOLDER") ; exit 1; fi
+# echo "---------"
+# echo "using REMOTE_USER: $REMOTE_USER"
+# echo "using REMOTE_HOST: $REMOTE_HOST"
+# echo "using SECRET_KEY: $SECRET_KEY"
+# echo "using BACKUP_FOLDER: $BACKUP_FOLDER"
+# echo "---------"
 REMOTE_USERHOST="$REMOTE_USER"@"$REMOTE_HOST"
-
 
 # checks
 if [ ! -r "$SECRET_KEY" ]; then
-	(>&2 echo " - (FAIL) ssh private key file not found or not readable: '$SECRET_KEY'.")
+	(>&2 echo " >>> (FAIL) ssh private key file not found or not readable: '$SECRET_KEY'.")
 	exit 1 
 fi
 echo " - (OK) found ssh private key file: '$SECRET_KEY'."
@@ -46,7 +47,7 @@ echo " - (OK) found ssh private key file: '$SECRET_KEY'."
 # ssh connection
 ssh -i "$SECRET_KEY" "$REMOTE_USERHOST" -q exit
 if [ $? -ne 0 ]; then
-	(>&2 echo " - (FAIL) there was a problem while trying to stablish an ssh connection to the remote server '$REMOTE_USERHOST'.")
+	(>&2 echo " >>> (FAIL) there was a problem while trying to stablish an ssh connection to the remote server '$REMOTE_USERHOST'.")
 	exit 1
 fi
 echo " - (OK) ssh connection works."
@@ -55,6 +56,10 @@ echo " - (OK) ssh connection works."
 # ssh script
 TEST_SCRIPT="/tmp/test_android_requests_backups_ssh.bash"
 rm -f "$TEST_SCRIPT"
+if [ -e "$TEST_SCRIPT" ]; then
+	(>&2 echo " >>> (FAIL) Failed to delete previous tmp file: '$TEST_SCRIPT'.")
+	exit 1 
+fi
 echo "mkdir -p \"$BACKUP_FOLDER\""                             >> "$TEST_SCRIPT"
 echo "if [ ! -d \"$BACKUP_FOLDER\" ]; then exit 1; fi"         >> "$TEST_SCRIPT"
 echo "rm -f \"$BACKUP_FOLDER\"/test_ssh_file"                  >> "$TEST_SCRIPT"
@@ -64,15 +69,11 @@ echo "if [ ! -w \"$BACKUP_FOLDER\" ]; then exit 1; exit 0; fi" >> "$TEST_SCRIPT"
 echo "exit 0"                                                  >> "$TEST_SCRIPT"
 ssh -i "$SECRET_KEY" "$REMOTE_USERHOST" "bash -s" -- < "$TEST_SCRIPT"
 if [ $? -ne 0 ]; then
-	(>&2 echo " - (FAIL): SSH script exited  with non zero status. Maybe, the '$REMOTE_USER' user does not has the proper permissions to create the folder: '$BACKUP_FOLDER'.")
+	(>&2 echo " >>> (FAIL): SSH script exited  with non zero status. Maybe, the '$REMOTE_USER' user does not has the proper permissions to create the folder: '$BACKUP_FOLDER'.")
 	exit 1
 fi
 echo " - (OK) ssh script worked ok. successfully connected and ran script on remote."
 
 # end
-echo " - (WIN) The SSH test to a remote host passed successfully."
-echo "################################################################################"
-echo ""
+echo " >>> (SUCCESS) The SSH test to a remote host passed successfully."
 exit 0
-#############################################################################################################
-
